User Releated workouts and auth by user

+++ frontend/src/components/WorkoutDetails.js
@@ -1,13 +1,21 @@
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext"
 //DateFns 
 import { formatDistanceToNow } from 'date-fns'
+import { useAuthContext } from "../hooks/useAuthContext"
 
 const WorkoutDetails = ({ workout }) => {
     const { dispatch } = useWorkoutsContext()
+    const { user } = useAuthContext()
 
     const handleClick = async () => {
+        // if (!user) return because the user is not logged in
+        if (!user) return
+
         const response = await fetch(`/api/workouts/${workout._id}` , {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers : {
+                'Authorization': `Bearer ${user.token}`
+            }
         })
         const json = await response.json()
